Unit & Integration test for individiual componet and API

+++ app/Http/Controllers/TaskController.php
@@ -14,14 +14,28 @@ class TaskController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $tasks = Task::where('user_id', Auth::id())
+        $userId = Auth::id();
+
+        // CHECK SCOPE (If using API Token)
+        // If the user is logged in via Token, ensure they have 'read' scope
+        if ($request->user()->tokenCan('tasks:read') === false) {
+            abort(403, 'Token does not have read permissions.');
+        }
+
+        $tasks = Task::where('user_id', $userId)
         ->with('category')
         ->latest()
         ->get();
 
-        $categories = Category::where('user_id', Auth::id())->get();
+        //return json for api, view for browser
+        if($request->wantsJson()){
+            return response()->json($tasks);
+        }
+
+        //For web view dashoard
+        $categories = Category::where('user_id', $userId)->get();
 
         return view ('dashboard', compact('tasks', 'categories'));
     }
@@ -48,6 +62,10 @@ class TaskController extends Controller
         if(config('services.slack.webhook_url')){
             Notification::route('slack', config('services.slack.webhook_url'))
                         ->notify(new TaskCreated($task));
+        }
+
+        if($request->wantsJson()){
+            return response()->json($task, 201);
         }
 
         return redirect()->route('dashboard')->with('success', 'Task Created!');
@@ -82,15 +100,24 @@ class TaskController extends Controller
 
         $task->update($validated);
 
+        if($request->wantsJson()){
+            return response()->json($task, 200);
+        }
+
         return response()->json($task->load('category'), 200);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Task $task)
+    public function destroy(Request $request, Task $task)
     {
         abort_if($task->user_id !== Auth::id(), 403, 'Unauthorized action.');
+
+        // CHECK SCOPE: Does this token have permission to write/delete?
+        if ($request->user()->tokenCan('tasks:write') === false) {
+            abort(403, 'Token is read-only.');
+        }
 
         $task->delete();
